# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Deriv Trading Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /strategies:
    get:
      operationId: listStrategies
      tags: [strategies]
      summary: List all strategies
      responses:
        "200":
          description: Strategies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Strategy"

    post:
      operationId: createStrategy
      tags: [strategies]
      summary: Create a new strategy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyInput"
      responses:
        "201":
          description: Strategy created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
  /strategies/{id}:
    get:
      operationId: getStrategy
      tags: [strategies]
      summary: Get a specific strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Strategy details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"
        "404":
          description: Not found

    patch:
      operationId: updateStrategy
      tags: [strategies]
      summary: Update a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StrategyUpdate"
      responses:
        "200":
          description: Updated strategy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Strategy"

    delete:
      operationId: deleteStrategy
      tags: [strategies]
      summary: Delete a strategy
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Indicators ──────────────────────────────────────────────────────────────
  /indicators:
    get:
      operationId: listIndicators
      tags: [indicators]
      summary: List all custom indicators
      responses:
        "200":
          description: Indicators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Indicator"

    post:
      operationId: createIndicator
      tags: [indicators]
      summary: Create a custom indicator
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndicatorInput"
      responses:
        "201":
          description: Indicator created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicator"
  /indicators/{id}:
    get:
      operationId: getIndicator
      tags: [indicators]
      summary: Get a specific indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Indicator details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicator"

    patch:
      operationId: updateIndicator
      tags: [indicators]
      summary: Update an indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IndicatorUpdate"
      responses:
        "200":
          description: Updated indicator
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Indicator"

    delete:
      operationId: deleteIndicator
      tags: [indicators]
      summary: Delete an indicator
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── Datasets ────────────────────────────────────────────────────────────────
...
  /datasets/backtests/upload:
    post:
      operationId: uploadBacktestDataset
      tags: [datasets]
      summary: Upload a new CSV dataset for backtesting
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/UploadDatasetInput"
      responses:
        "200":
          description: Dataset uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  filename:
                    type: string
                  success:
                    type: boolean
                required:
                  - filename
                  - success

  # ─── Backtests ───────────────────────────────────────────────────────────────
  /backtests:
    get:
      operationId: listBacktests
      tags: [backtests]
      summary: List all backtest runs
      parameters:
        - name: strategyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Backtests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Backtest"

    post:
      operationId: runBacktest
      tags: [backtests]
      summary: Run a strategy backtest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BacktestInput"
      responses:
        "201":
          description: Backtest started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backtest"
  /strategies/{id}/webhook/test:
    post:
      operationId: testStrategyWebhook
      tags: [strategies]
      summary: Send a sample signal payload to the strategy webhook URL
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Webhook test result
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  status:
                    type: ["integer", "null"]
                  error:
                    type: ["string", "null"]
                required: [ok]
  /backtests/{id}:
    get:
      operationId: getBacktest
      tags: [backtests]
      summary: Get backtest result
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Backtest result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Backtest"
    delete:
      operationId: deleteBacktest
      tags: [backtests]
      summary: Delete a backtest
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Backtest deleted

  # ─── AI Analysis ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Strategy:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [vanilla_options, forex, multiplier, universal]
        code:
          type: string
        parameters:
          type: ["string", "null"]
        isActive:
          type: boolean
        isPublic:
          type: boolean
        winRate:
          type: ["number", "null"]
        totalBacktests:
          type: integer
        webhookUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - type
        - code
        - isActive
        - isPublic
        - totalBacktests
        - createdAt
        - updatedAt
    StrategyInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [vanilla_options, forex, multiplier, universal]
        code:
          type: string
        parameters:
          type: ["string", "null"]
        isPublic:
          type: boolean
        webhookUrl:
          type: ["string", "null"]
      required:
        - name
        - type
        - code
    StrategyUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        type:
          type: string
          enum: [vanilla_options, forex, multiplier, universal]
        code:
          type: string
        parameters:
          type: ["string", "null"]
        isActive:
          type: boolean
        isPublic:
          type: boolean
        webhookUrl:
          type: ["string", "null"]

    # ─── Indicators ─────────────────────────────────────────────────────────────
    Indicator:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: string
        name:
          type: string
        description:
          type: ["string", "null"]
        code:
          type: string
        parameters:
          type: ["string", "null"]
        isPublic:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - code
        - isPublic
        - createdAt
        - updatedAt
    IndicatorInput:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        code:
          type: string
        parameters:
          type: ["string", "null"]
        isPublic:
          type: boolean
      required:
        - name
        - code
    IndicatorUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        code:
          type: string
        parameters:
          type: ["string", "null"]
        isPublic:
          type: boolean

    # ─── Backtests ──────────────────────────────────────────────────────────────
    Backtest:
      type: object
      properties:
        id:
          type: integer
        strategyId:
          type: integer
        userId:
          type: string
        symbol:
          type: string
        fromDate:
          type: string
          format: date-time
        toDate:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, running, completed, failed]
        totalTrades:
          type: ["integer", "null"]
        winRate:
          type: ["number", "null"]
        totalPnl:
          type: ["number", "null"]
        maxDrawdown:
          type: ["number", "null"]
        sharpeRatio:
          type: ["number", "null"]
        results:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: ["string", "null"]
          format: date-time
      required:
        - id
        - strategyId
        - userId
        - symbol
        - fromDate
        - toDate
        - status
        - createdAt
...
    UploadDatasetInput:
      type: object
      properties:
        file:
          type: string
          format: binary
      required:
        - file

    # ─── AI Analysis ────────────────────────────────────────────────────────────
...